test helper and simplecov
