+++ .gitignore
@@ -1,2 +1,3 @@
 Gemfile.lock
 capybara*.html
+coverage
