+++ test/test_helper.rb
@@ -1,3 +1,6 @@
+require 'simplecov'
+SimpleCov.start
+
 ENV["RACK_ENV"] ||= "test"
 
 require 'bundler'
@@ -9,9 +12,28 @@ require 'minitest/pride'
 require 'capybara'
 require 'pry'
 
-Capybara.app = TrafficSpy::Server
-
 DatabaseCleaner.strategy = :truncation, {except: %w[public.schema_migrations]}
+
+class Minitest::Test
+
+  def setup
+    DatabaseCleaner.start
+  end
+
+  def teardown
+    DatabaseCleaner.clean
+  end
+end
+
+class ControllerTest < Minitest::Test
+  include Rack::Test::Methods
+
+  def app
+    TrafficSpy::Server
+  end
+end
+
+Capybara.app = TrafficSpy::Server
 
 class FeatureTest < Minitest::Test
   include Capybara::DSL
